Own module for typed signals + variadic args
